one controller for assets and slots, code optimisation
