+++ cartridges/custom_cartridge_1/cartridge/controllers/AssetsAndSlots.js
@@ -1,19 +1,18 @@
 "use strict";
 
 /**
- * @namespace LoggedOrGuestUser
+ * @namespace AssetsAndSlots
  */
 
 var server = require("server");
 var cache = require("*/cartridge/scripts/middleware/cache");
 var consentTracking = require("*/cartridge/scripts/middleware/consentTracking");
 var pageMetaData = require("*/cartridge/scripts/middleware/pageMetaData");
+
 /**
- * 
-/**
- * @name Base/LoggedOrGuestUser-Show
+ * @name Base/AssetsAndSlots-Slot
  * @function
- * @memberof LoggedOrGuestUser
+ * @memberof AssetsAndSlots
  * @param {middleware} - consentTracking.consent
  * @param {middleware} - cache.applyDefaultCache
  * @param {category} - non-sensitive
@@ -21,7 +20,34 @@ var pageMetaData = require("*/cartridge/scripts/middleware/pageMetaData");
  * @param {serverfunction} - get
  */
 server.get(
-    "Show",
+    "Slot",
+    consentTracking.consent,
+    cache.applyDefaultCache,
+    function (req, res, next) {
+        var Site = require("dw/system/Site");
+        var pageMetaHelper = require("*/cartridge/scripts/helpers/pageMetaHelper");
+
+        pageMetaHelper.setPageMetaTags(req.pageMetaData, Site.current);
+
+        res.render("slotTemplate");
+
+        next();
+    },
+    pageMetaData.computedPageMetaData
+);
+
+/**
+ * @name Base/AssetsAndSlots-Assets
+ * @function
+ * @memberof AssetsAndSlots
+ * @param {middleware} - consentTracking.consent
+ * @param {middleware} - cache.applyDefaultCache
+ * @param {category} - non-sensitive
+ * @param {renders} - isml
+ * @param {serverfunction} - get
+ */
+server.get(
+    "Assets",
     consentTracking.consent,
     cache.applyDefaultCache,
     function (req, res, next) {
@@ -32,19 +58,19 @@ server.get(
 
         pageMetaHelper.setPageMetaTags(req.pageMetaData, Site.current);
 
-        var loggedInAsset = ContentMgr.getContent('Logged');
-        var guestAsset = ContentMgr.getContent('Guest');
-        var loggedInAssetBody = loggedInAsset.custom.body;
-        var guestAssetBody = guestAsset.custom.body;
 
         var currentCustomer = req.currentCustomer.profile;
         var messageLoogedOrGuest;
 
         if (currentCustomer) {
+            var loggedInAsset = ContentMgr.getContent('Logged');
+            var loggedInAssetBody = loggedInAsset.custom.body;
             var customerName = currentCustomer.firstName;
-            var loggedInAssetBodyAsString = loggedInAssetBody + "";
+            var loggedInAssetBodyAsString = loggedInAssetBody .toString();
             messageLoogedOrGuest = loggedInAssetBodyAsString.replace("{0}", customerName);
         } else {
+            var guestAsset = ContentMgr.getContent('Guest');
+            var guestAssetBody = guestAsset.custom.body;
             messageLoogedOrGuest = guestAssetBody;
         }
 
@@ -53,7 +79,6 @@ server.get(
         next();
     },
     pageMetaData.computedPageMetaData
-
 );
 
 module.exports = server.exports();
