--- cartridges/custom_cartridge_1/cartridge/controllers/SlotDemo.js
@@ -1,42 +1,0 @@
-"use strict";
-
-/**
- * @namespace SlotDemoDemo
- */
-
-var server = require("server");
-var cache = require("*/cartridge/scripts/middleware/cache");
-var consentTracking = require("*/cartridge/scripts/middleware/consentTracking");
-var pageMetaData = require("*/cartridge/scripts/middleware/pageMetaData");
-
-
-/**
- * Any customization on this endpoint, also requiresss update for Default-Start endpoint
- */
-/**
- * @function
- * @memberof SlotDemo
- * @param {middleware} - consentTracking.consent
- * @param {middleware} - cache.applyDefaultCache
- * @param {category} - non-sensitive
- * @param {renders} - isml
- * @param {serverfunction} - get
- */
-server.get(
-    "Show",
-    consentTracking.consent,
-    cache.applyDefaultCache,
-    function (req, res, next) {
-        var Site = require("dw/system/Site");
-        var pageMetaHelper = require("*/cartridge/scripts/helpers/pageMetaHelper");
-
-        pageMetaHelper.setPageMetaTags(req.pageMetaData, Site.current);
-
-        res.render("slotTemplate");
-       
-        next();
-    },
-    pageMetaData.computedPageMetaData
-);
-
-module.exports = server.exports();
